Add windows support to Native driver by breaking up platform specific data and behavior using the standard mechanisms of go build pragmas, embedded structs, and shared functions
